requirements.txt
streamlit
openai
python-dotenv
reportlab
pillow
PyMuPDF
reportlab==4.2.5